refactor: adjust table APIs to skip passing options
