+++ crates/core/src/lib.rs
@@ -17,14 +17,7 @@
  * under the License.
  */
 
-use crate::table::Table;
-
-pub mod file_group;
-pub mod table;
-pub type HudiTable = Table;
 pub mod config;
+pub mod file_group;
 pub mod storage;
-
-pub fn crate_version() -> &'static str {
-    env!("CARGO_PKG_VERSION")
-}
+pub mod table;
