+++ python/hudi/__init__.py
@@ -15,7 +15,5 @@
 #  specific language governing permissions and limitations
 #  under the License.
 
-from ._internal import __version__ as __version__
-from ._internal import rust_core_version as rust_core_version
 from ._internal import HudiFileSlice as HudiFileSlice
-from .table import HudiTable as HudiTable
+from ._internal import HudiTable as HudiTable
