+++ python/src/_file_group.rs
@@ -1,0 +1,76 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+use std::path::PathBuf;
+
+use anyhow::Context;
+use pyo3::{pyclass, pymethods, PyErr, PyResult};
+
+use hudi::file_group::FileSlice;
+
+#[cfg(not(tarpaulin))]
+#[derive(Clone, Debug)]
+#[pyclass]
+pub struct HudiFileSlice {
+    #[pyo3(get)]
+    file_group_id: String,
+    #[pyo3(get)]
+    partition_path: String,
+    #[pyo3(get)]
+    commit_time: String,
+    #[pyo3(get)]
+    base_file_name: String,
+    #[pyo3(get)]
+    base_file_size: usize,
+    #[pyo3(get)]
+    num_records: i64,
+}
+
+#[cfg(not(tarpaulin))]
+#[pymethods]
+impl HudiFileSlice {
+    pub fn base_file_relative_path(&self) -> PyResult<String> {
+        PathBuf::from(&self.partition_path)
+            .join(&self.base_file_name)
+            .to_str()
+            .map(String::from)
+            .context(format!(
+                "Failed to get base file relative path for file slice: {:?}",
+                self
+            ))
+            .map_err(PyErr::from)
+    }
+}
+
+#[cfg(not(tarpaulin))]
+pub fn convert_file_slice(f: &FileSlice) -> HudiFileSlice {
+    let file_group_id = f.file_group_id().to_string();
+    let partition_path = f.partition_path.as_deref().unwrap_or_default().to_string();
+    let commit_time = f.base_file.commit_time.to_string();
+    let base_file_name = f.base_file.info.name.clone();
+    let base_file_size = f.base_file.info.size;
+    let num_records = f.base_file.stats.clone().unwrap_or_default().num_records;
+    HudiFileSlice {
+        file_group_id,
+        partition_path,
+        commit_time,
+        base_file_name,
+        base_file_size,
+        num_records,
+    }
+}
